Improve unit tests so each file can be run by itself and still pass.
