+++ lib/chef/http/socketless_chef_zero_client.rb
@@ -43,6 +43,7 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+require "net/http"
 require "chef_zero/server"
 require_relative "../dist"
