+++ spec/support/shared/unit/testable_shell_session.rb
@@ -1,0 +1,21 @@
+require "chef/shell/shell_session"
+
+class TestableShellSession < Shell::ShellSession
+
+  def rebuild_node
+    nil
+  end
+
+  def rebuild_collection
+    nil
+  end
+
+  def loading
+    nil
+  end
+
+  def loading_complete
+    nil
+  end
+
+end
