+++ spec/unit/server_api_spec.rb
@@ -1,4 +1,5 @@
 require "spec_helper"
+require "support/shared/unit/versioned_class_versions"
 
 SIGNING_KEY_DOT_PEM = "-----BEGIN RSA PRIVATE KEY-----
 MIIEpAIBAAKCAQEA49TA0y81ps0zxkOpmf5V4/c4IeR5yVyQFpX3JpxO4TquwnRh
@@ -59,22 +60,6 @@ describe Chef::ServerAPI do
   end
 
   context "versioned apis" do
-    class VersionedClassV0
-      extend Chef::Mixin::VersionedAPI
-      minimum_api_version 0
-    end
-
-    class VersionedClassV2
-      extend Chef::Mixin::VersionedAPI
-      minimum_api_version 2
-    end
-
-    class VersionedClassVersions
-      extend Chef::Mixin::VersionedAPIFactory
-      add_versioned_api_class VersionedClassV0
-      add_versioned_api_class VersionedClassV2
-    end
-
     before do
       Chef::ServerAPIVersions.instance.reset!
     end
