+++ spec/unit/shell_spec.rb
@@ -17,25 +17,7 @@
 
 require "spec_helper"
 require "ostruct"
-
-ObjectTestHarness = Proc.new do
-  extend Shell::Extensions::ObjectCoreExtensions
-
-  def conf=(new_conf)
-    @conf = new_conf
-  end
-
-  def conf
-    @conf
-  end
-
-  desc "rspecin'"
-  def rspec_method; end
-end
-
-class TestJobManager
-  attr_accessor :jobs
-end
+require "support/shared/unit/test_job_manager.rb"
 
 describe Shell do
